Remove warning of unused static function
